+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This major release introduces a few breaking changes. As such, we've provided an
 - Added new LR schedulers: `LinearWithWarmup`, `InvSqrtWithWarmup`, `ConstantWithWarmup`, `SequentialScheduler`.
 - Added option to pre-download checkpoint files from remote storage before trying to load a checkpoint.
 - Added a callback for sending Slack notifications.
+- The trainer can load model-only checkpoints now.
 
 ### Changed